Fixes #24.
A certain set of functions that exist both in batch and standalone mode
are dynamically bound as local names in the functions subpackage. That makes
them undiscoverable by IDEs like PyCharm. Adding a straightforward import
statement for these function names -- even though the names are re-bound
immediately with wrapped standalone versions of the functions -- makes
the function names visible to PyCharm.